+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Changelog
 
-## Unreleased
+## 0.19.0
+
+### Features
+
+- Add rich context info to native crashes ([#747](https://github.com/getsentry/sentry-unity/pull/747))
+- Include build ID in an event release info ([#795](https://github.com/getsentry/sentry-unity/pull/795))
+
+### Fixes
+
+- Don't report Aplication-Not-Responding while the app is in the background ([#796](https://github.com/getsentry/sentry-unity/pull/796))
+
+## 0.18.0
 
 ### Features